pool代码

+++ contracts/wtfswap/interfaces/IPool.sol
@@ -69,8 +69,8 @@ interface IPool{
     event Burn(
         address indexed owner,
         uint128 amount,
-        uint128 amount0,
-        uint128 amount1
+        uint256 amount0,
+        uint256 amount1
     );
     //移除流动性
     function burn(
